[feature/RPA-681]: Reusable carousel mechanism
