+++ runbotics/runbotics-orchestrator-ui/src/src-landing/components/Carousel/Carousel.types.ts
@@ -1,0 +1,7 @@
+export interface CarouselSlide {
+    content: JSX.Element;
+}
+
+export interface CarouselProps {
+    slides: CarouselSlide[];
+}
