+++ runbotics/runbotics-orchestrator-ui/src/src-landing/components/Carousel/MobileCarouselNav/MobileCarouselNav.module.scss
@@ -5,11 +5,11 @@
 }
 
 .counter {
-    composes: counter from '../../views/sections/BenefitsSection/BenefitsCarousel/BenefitsCarousel.module.scss';
+    composes: counter from '../Carousel.module.scss';
 }
 
 .counterCurrent {
-    composes: counterCurrent from '../../views/sections/BenefitsSection/BenefitsCarousel/BenefitsCarousel.module.scss';
+    composes: counterCurrent from '../Carousel.module.scss';
 }
 
 @media (max-width: $breakpoint-sm) {
