+++ runbotics/runbotics-orchestrator-ui/src/src-landing/components/Carousel/MobileCarouselNav/MobileCarouselNav.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import CarouselButton from '#src-landing/components/CarouselButton';
+import CarouselButton from '#src-landing/components/Carousel/CarouselButton';
 import Typography from '#src-landing/components/Typography';
 
 import styles from './MobileCarouselNav.module.scss';
