+++ runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/BenefitsSection/BenefitsArticle/BenefitsArticle.types.ts
@@ -1,0 +1,10 @@
+import benefitsTranslation from '#src-landing/translations/en/benefits.json';
+
+type ArticleTranslationKey = keyof typeof benefitsTranslation;
+
+export interface Article {
+    titleKey: ArticleTranslationKey;
+    descriptionKey: ArticleTranslationKey;
+    imageSrc: string;
+    imageAlt: string;
+}
