+++ runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/BenefitsSection/BenefitsArticle/index.ts
@@ -1,0 +1,3 @@
+import BenefitsArticle from './BenefitsArticle';
+
+export default BenefitsArticle;
