--- runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/BenefitsSection/BenefitsCarousel/BenefitsCarousel.types.ts
@@ -1,10 +1,0 @@
-import benefitsTranslation from '#src-landing/translations/en/benefits.json';
-
-type SlideTranslationKey = keyof typeof benefitsTranslation;
-
-export interface Slide {
-    titleKey: SlideTranslationKey;
-    descriptionKey: SlideTranslationKey;
-    imageSrc: string;
-    imageAlt: string;
-}
