--- runbotics/runbotics-orchestrator-ui/src/src-landing/views/sections/BenefitsSection/BenefitsCarousel/index.ts
@@ -1,3 +1,0 @@
-import BenefitsCarousel from './BenefitsCarousel';
-
-export default BenefitsCarousel;
